Develop

+++ src/main/java/com/twoxplusone/rshardware/rs_hardware_glass_and_electrical/service/InvoiceService.java
@@ -2,11 +2,16 @@ package com.twoxplusone.rshardware.rs_hardware_glass_and_electrical.service;
 
 import com.twoxplusone.rshardware.rs_hardware_glass_and_electrical.Entity.CustomerInvoices;
 import com.twoxplusone.rshardware.rs_hardware_glass_and_electrical.Repository.InvoiceRepositoy;
+import com.twoxplusone.rshardware.rs_hardware_glass_and_electrical.dtos.ApiResponse;
+import com.twoxplusone.rshardware.rs_hardware_glass_and_electrical.liveconfig.Payment;
 import lombok.extern.slf4j.Slf4j;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
+import org.togglz.core.manager.FeatureManager;
 
 import java.time.LocalDate;
 import java.util.Comparator;
@@ -18,14 +23,24 @@ public class InvoiceService {
     @Autowired
     private InvoiceRepositoy invoiceRepositoy;
 
+    @Autowired
+    private FeatureManager featureManager;
 
-    public CustomerInvoices createInvoice(CustomerInvoices customerInvoices){
-        CustomerInvoices saveInvoice = invoiceRepositoy.save(customerInvoices);
-        return saveInvoice;
-
+    public ResponseEntity<?> createInvoice(CustomerInvoices customerInvoices){
+        if(featureManager.isActive(Payment.PAYMENT_RECEIVED)){
+            CustomerInvoices saveInvoice = invoiceRepositoy.save(customerInvoices);
+            return ResponseEntity.ok(saveInvoice);
+        }else {
+            ApiResponse serverNotRenewed = ApiResponse.builder()
+                    .message("Server Not Renewed, Contact support.")
+                    .success(false)
+                    .build();
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(serverNotRenewed);
+        }
     }
 
     public List<CustomerInvoices> getAllInvoices() {
+        if(featureManager.isActive(Payment.PAYMENT_RECEIVED)){
         List<CustomerInvoices> allInvoices = invoiceRepositoy.findAll();
 
         return allInvoices.stream()
@@ -44,15 +59,26 @@ public class InvoiceService {
                     }
                 }).reversed())
                 .toList();
+        }else {
+            return List.of();
+        }
     }
 
     public CustomerInvoices fetchInvoiceById(String id){
-        CustomerInvoices customerInvoices = invoiceRepositoy.findById(id).orElse(null);
-        return customerInvoices;
+        if(featureManager.isActive(Payment.PAYMENT_RECEIVED)){
+            CustomerInvoices customerInvoices = invoiceRepositoy.findById(id).orElse(null);
+            return customerInvoices;
+        }else{
+            return null;
+        }
+
     }
 
     public void deleteInvoiceById(String id){
-        invoiceRepositoy.deleteById(id);
+        if(featureManager.isActive(Payment.PAYMENT_RECEIVED)){
+            invoiceRepositoy.deleteById(id);
+        }
+
     }
 
 }
